Changing points types to decimal.

+++ src/Entelect.Encentivize.Sdk/Achievements/AchievementInput.cs
@@ -5,7 +5,7 @@ namespace Entelect.Encentivize.Sdk.Achievements
     public class AchievementInput
     {
         public long AchievementId { get; set; }
-        public int? OverriddenPoints { get; set; }
+        public decimal? OverriddenPoints { get; set; }
         public DateTime DateAchievementEarnedUtc { get; set; }
         public string Comment { get; set; }
     }
